requirements.txt
django 
djangorestframework 
django-filter
djangorestframework-simplejwt 
psycopg2-binary 
python-decouple 
Pillow 
Celery 
redis
drf-yasg
docker
unicorn